+++ src/Database/Migrations/2021_06_27_043322_add_alias_column_to_roles_and_permissions_tables.php
@@ -1,0 +1,38 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+class AddAliasColumnToRolesAndPermissionsTables extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::table('roles', function (Blueprint $table) {
+            $table->string('alias', 190)->nullable();
+        });
+        Schema::table('permissions', function (Blueprint $table) {
+            $table->string('alias', 190)->nullable();
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::table('roles', function (Blueprint $table) {
+            $table->dropColumn('alias');
+        });
+        Schema::table('permissions', function (Blueprint $table) {
+            $table->dropColumn('alias');
+        });
+    }
+}