+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.1.3.1] - 2021-06-27
+
+### Added
+
+- Translatable alias for Roles and Permissions
+
+### Changed
+
+### Fixed
+
+- Seeders debugged
+
+
 ## [1.1.3] - 2021-06-26
 
 ### Added